dashboards now live in a top level directory
