--- server/middleware/dashboards/anonymous.yml
@@ -1,39 +1,0 @@
----
-id: anonymous
-title: Anonymous
-isprimary: true
-description: <dt>"Though it's cold and lonely in the deep dark night, I can see paradise by the dashboard light"</dt><dd>— Meat Loaf</dd>
-charts:
-  -
-    question: "Are daily anonymous visitor counts increasing?"
-    dashboardfeature: "anonymous"
-    queryname: "daily"
-    name: "anonymous/daily"
-    query: "page:view->count(user.ab.allocationId)->filter(!user.uuid)"
-    datalabel: "Anonymous visitors"
-  -
-    question: "What sites are anons visiting from?"
-    dashboardfeature: "anonymous"
-    queryname: "referrers"
-    name: "anonymous/referrers"
-    query: "page:view->count(user.ab.allocationId)->filter(!user.uuid)->group(page.referrer.host)"
-    printer: "PieChart"
-    datalabel: "Continent"
-  -
-    question: "Are anons clicking on the propensity AB test"
-    dashboardfeature: "anonymous"
-    queryname: "propensityAB/clicks"
-    name: "anonymous/propensityAB/clicks"
-    query: "cta:click->count()->filter(context.domPath~propensity-discount)->group(user.ab.tests.propensityAbTest)->tidy()"
-  -
-    question: "Are anons in the AB test viewing more pages"
-    dashboardfeature: "anonymous"
-    queryname: "propensityAB/page-views"
-    name: "anonymous/propensityAB/page-views"
-    query: "page:view->count()->group(user.ab.tests.propensityAbTest)->tidy()"
-  -
-    question: "Are anons in the AB test viewing more pages"
-    dashboardfeature: "anonymous"
-    queryname: "propensityAB/page-views"
-    name: "anonymous/propensityAB/page-views"
-    query: "page:view->count()->group(user.ab.tests.propensityAbTest)->tidy()"
