--- server/middleware/dashboards/optinout.yml
@@ -1,39 +1,0 @@
----
-id: optinout
-title: Opt in/out
-isprimary: true
-description: <dt>"Though it's cold and lonely in the deep dark night, I can see paradise by the dashboard light"</dt><dd>— Meat Loaf</dd>
-charts:
-  -
-    question: "What percentage of users opt out"
-    dashboardfeature: "optinout"
-    queryname: "opt-out-rate"
-    name: "optinout/opt-out-rate"
-    query: "@pct(site:optout->count(user.uuid),page:view->count(user.uuid))->round()"
-  -
-    question: "What percentage of users opt in"
-    dashboardfeature: "optinout"
-    queryname: "opt-in-rate"
-    name: "optinout/opt-in-rate"
-    query: "@pct(site:optin->count(user.uuid)->filter(context.optedVia!=__anon-opt-in)->filter(context.optedVia!=__m-dot-opt-in),page:view->count(user.uuid))->round()"
-  -
-    question: "Number of users who opted out"
-    dashboardfeature: "optinout"
-    queryname: "total-optouts"
-    name: "optinout/total-optouts"
-    query: "site:optout->count(user.uuid)"
-    printer: "ColumnChart"
-  -
-    question: "Number of users who opted in"
-    dashboardfeature: "optinout"
-    queryname: "total-optins"
-    name: "optinout/total-optins"
-    query: "site:optin->count(user.uuid)->filter(context.optedVia!=__anon-opt-in)->filter(context.optedVia!=__m-dot-opt-in)"
-    printer: "ColumnChart"
-  -
-    question: "Number of users who used the site"
-    dashboardfeature: "optinout"
-    queryname: "total-users"
-    name: "optinout/total-users"
-    query: "page:view->count(user.uuid)"
-    printer: "ColumnChart"
