--- server/middleware/dashboards/overview.yml
@@ -1,21 +1,0 @@
----
-id: overview
-title: nextFT overview
-isprimary: false
-description: <dt>"Though it's cold and lonely in the deep dark night, I can see paradise by the dashboard light"</dt><dd>— Meat Loaf</dd>
-charts:
-  -
-    question: "What articles have been viewed in the last 30 minutes?"
-    dashboardfeature: "overview"
-    queryname: "popular"
-    name: "overview/popular"
-    query: "page:view->count(user.uuid)->relTime(this_30_minutes)->group(content.title)->tidy()"
-    datalabel: "Articles"
-    printer: "PieChart"
-  -
-    question: "How many members do we have?"
-    dashboardfeature: "overview"
-    queryname: "users"
-    name: "overview/users"
-    query: "page:view->count(user.uuid)"
-    datalabel: "Members"
