--- server/middleware/dashboards/performance.yml
@@ -1,41 +1,0 @@
----
-id: performance
-title: Performance
-isprimary: true
-description: Fastest site in the west
-charts:
-  -
-    question: "timing of loadEventStart"
-    dashboardfeature: "performance"
-    queryname: "loadEventStart"
-    name: "performance/loadEventStart"
-    query: "page:load-timing->med(context.timings.offset.loadEventStart)"
-    datalabel: loadEventStart
-  -
-    question: "timing of domComplete"
-    dashboardfeature: "performance"
-    queryname: "domComplete"
-    name: "performance/domComplete"
-    query: "page:load-timing->med(context.timings.offset.domComplete)"
-    datalabel: domComplete
-  -
-    question: "timing of domContentLoadedEventStart"
-    dashboardfeature: "performance"
-    queryname: "domContentLoadedEventStart"
-    name: "performance/domContentLoadedEventStart"
-    query: "page:load-timing->med(context.timings.offset.domContentLoadedEventStart)"
-    datalabel: domContentLoadedEventStart
-  -
-    question: "timing of domInteractive"
-    dashboardfeature: "performance"
-    queryname: "domInteractive"
-    name: "performance/domInteractive"
-    query: "page:load-timing->med(context.timings.offset.domInteractive)"
-    datalabel: domInteractive
-  -
-    question: "timing of fontsLoaded"
-    dashboardfeature: "performance"
-    queryname: "fontsLoaded"
-    name: "performance/fontsLoaded"
-    query: "page:load-timing->med(context.timings.marks.fontsLoaded)"
-    datalabel: fontsLoaded
